New verb: rightandwrong
When used, it will arm every human mob that is not already a special role (and
with a client assigned) with a random gun from the following list:
-Taser
-Egun
-Laser
-Revolver
-C20r
-Decloner
-Desert Eagle
TG: -Gyrojet Pistol
-Pulse Rifle
-Silenced Pistol
-Laser Cannon
-Combat Shotgun
-Mini-Uzi
-Mini-Ebow

It will then inform the player that
"You are the survivor! Your own safety matters above all else, trust no one and
kill anyone who gets in your way. However, armed as you are, now would be the
perfect time to settle that score or grab that pair of yellow gloves you've been
eyeing..."

And make them a traitor/give them the objective to survive till the round ends.

After use, rightandwrong will remove itself from the users list of verbs.

rightandwrong addeded to the Spellbook under "Summon Guns"
The verb will appear in the spell tab.

Updated changelog
Revision: r3071
Author: 	 kortgstation