TG: - Moved three (well two, one was doubled for some reason) screen/attackby() procs to appear AFTER the screen object is defined. They were in a random file instead of where the objects are. If you double-click and 'screen' object in the object browser it now takes you to the definition instead of some random file.
- Clicking a grille with a glass or r-glass sheet in your hand, when the grille
is in one of your cardinal directions (North, south, east or west) will make you
start building a window. It takes 2s to build the window, which spawns
unfastened (same as when you create one.) This should make fixing damaged
grille-window combinations easier. If you are standing on a grille and click it
with a glass / r-glass sheet in your hand, the window will face the direction
you're currently facing.

Screenshot:
http://www.kamletos.si/placing%20windows%20on%20grilles.PNG
Revision: r3081
Author: 	 baloh.matevz