TG: Updated toilets, urinals, and showers with NEW features.
The shower path has changed!
From /obj/structure/shower to /obj/machinery/shower

Made bookcases appear empty if they have no books, and fill up as more books are
placed in them.

Updated do_after() to have an extra var that determines if the user is required
to keep the item in the active hand. This is only used in swirlies at the
moment.
Revision: r3684
Author: 	 petethegoat