Refined earlier fix
Tried a couple of other solutions, but this ended up being the most reliable.

Now checks all areas associated with the firedoor to see if they should still be down.

Also makes a final check before closing the door in case the fire alarm turns off in the middle of an autoclose cycle.