ZAS_Debug verb.
Applys a client side only layer of colors over the various zones to show if they are connected or not.

Green is your current zone
Blue is adjacent zones that are connected.
yellow is zones that are connected but not adjacent to your current zone.
Red is for zones that aren't connected at all.