Adds Mentors.
Adds a config option MENTORS which sets the variable config.mods_are_mentors
Adds a rights level of R_MENTOR which gets msay, private message, aghost, notes, and a new proc for checking for new players (requires database support).

If the confic option for mentors is set then

	the ckeys listed in moderators.txt file will instead be set as mentors, you can still make moderators by adding them in admins.txt

	staffwho will show Mentors instead of Moderators as the heading above the listing of non-admins.

Also:  Players now get a message gently reminding them to click the name of the staff member to reply instead of ahelping over and over.