Map fixes-
PsiOmegaDelta: "Atmospherics
Stacked manifolds at 91,86,1 and 93,86,1. The duplicates pop out at round-start.
The scrubber in Atmospherics begins off.

175,205,5 - Research outpost
Has a manifold and output vent attempting to connect into the same pipe.
The manifold used to be a simple bent pipe connecting the syphon vent with the omni-filter.

138,75,5 - Mining outpost
An UPA and air supply pipe share the same turf."