pAI submission process change
Attempts  to fix  #6481.

Previously it appears observers would be asked once, and only once, if they wished to become a pAI.
If one closed the configuration window without pressing Submit you'd never get the chance to apply again.

Now asks observers every time someone requests a pAI, similar to posibrains and Dionaea pods.
There is already a built-in ask delay of one minute between such requests and ghosts can decide to opt out for the rest of the round.